Moved examples and templates
